Refactoring: move GPTChatCompletionTemplate out of impl package since that's used in the API
